pyproject.toml
[project]
name = "marisa"
version = "0.1.0"
description = "Welcome to the world of cultivation."
authors = [
    { name = "KomoriDev", email = "mute231010@gmail.com" },
]
dependencies = [
    "nonebot2>=2.4.1",
    "nonebot-plugin-alconna>=0.54.2",
    "aiofiles>=24.1.0",
    "nonebot-plugin-orm>=0.7.6",
    "nonebot-plugin-userinfo>=0.2.4",
    "nonebot-plugin-waiter>=0.8.0",
    "nonebot-plugin-htmlrender>=0.5.1",
    "ruamel-yaml>=0.18.10",
    "nonebot-plugin-uninfo>=0.6.5",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/MyXiaoNan/Marisa"
repository = "https://github.com/MyXiaoNan/Marisa"

[project.optional-dependencies]
adapters = [
    "nonebot-adapter-onebot>=2.4.4",
    "nonebot-adapter-telegram>=0.1.0b17",
]
extra = [
  "nonebot2[fastapi]>=2.4.1",
  "nonebot-plugin-orm[aiosqlite]>=0.7.6",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["marisa"]

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.9.1",
    "isort>=5.13.2",
    "pre-commit>=4.0.1",
]

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.ruff.lint.isort]
length-sort = true
force-sort-within-sections = true
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "standard"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
